CS-296 Allow ReadFrame parse both client and server response frames
This commit enhances the ReadFrame utility funciton to receive and
process both frames recieved from the clients as well as those
received from the server at the backend. It first tries to parse the
just-read frame as a client request frame (the more common path) and
if it fails attempts to parse it as a server response frame.